fix: sonarqube errors

+++ PerfumeTracker.Server/Repo/PerfumeRepo.cs
@@ -1,21 +1,21 @@
 ﻿using Mapster;
 using Microsoft.EntityFrameworkCore;
-using PerfumeTracker.Server.DTO;
-using PerfumeTrackerAPI.DTO;
+using PerfumeTracker.Server.Dto;
+using PerfumeTrackerAPI.Dto;
 using PerfumeTrackerAPI.Models;
 using static PerfumeTrackerAPI.Repo.ResultType;
 
 namespace PerfumeTrackerAPI.Repo {
 	public class PerfumeRepo(PerfumetrackerContext context) {
-		public async Task<List<PerfumeWithWornStatsDTO>> GetPerfumesWithWorn(string fulltext = null) {
+		public async Task<List<PerfumeWithWornStatsDto>> GetPerfumesWithWorn(string? fulltext = null) {
 			var raw = await context
 				.Perfumes
 				.Where(p => string.IsNullOrWhiteSpace(fulltext)
 					|| p.FullText.Matches(EF.Functions.ToTsQuery($"{fulltext}:*"))
 					|| p.PerfumeTags.Any(pt => EF.Functions.ILike(pt.Tag.TagName, fulltext))
 					)
-				.Select(p => new PerfumeWithWornStatsDTO(
-					 new PerfumeDTO(
+				.Select(p => new PerfumeWithWornStatsDto(
+					 new PerfumeDto(
 						p.Id,
 						p.House,
 						p.PerfumeName,
@@ -27,7 +27,7 @@ namespace PerfumeTrackerAPI.Repo {
 						p.Spring,
 						p.Summer,
 						p.Winter,
-						p.PerfumeTags.Select(tag => new TagDTO(tag.Tag.TagName, tag.Tag.Color, tag.Tag.Id)).ToList()
+						p.PerfumeTags.Select(tag => new TagDto(tag.Tag.TagName, tag.Tag.Color, tag.Tag.Id)).ToList()
 					  ),
 					  p.PerfumeWorns.Any() ? p.PerfumeWorns.Count : 0,
 					  p.PerfumeWorns.Any() ? p.PerfumeWorns.Max(x => x.Created_At) : null
@@ -37,12 +37,12 @@ namespace PerfumeTrackerAPI.Repo {
 				.ToListAsync();
 			return raw;
 		}
-		public async Task<PerfumeWithWornStatsDTO?> GetPerfume(int id) {
+		public async Task<PerfumeWithWornStatsDto?> GetPerfume(int id) {
 			return await context
 				.Perfumes
 				.Where(p => p.Id == id)
-				.Select(p => new PerfumeWithWornStatsDTO(
-					  new PerfumeDTO(
+				.Select(p => new PerfumeWithWornStatsDto(
+					  new PerfumeDto(
 						p.Id,
 						p.House,
 						p.PerfumeName,
@@ -54,7 +54,7 @@ namespace PerfumeTrackerAPI.Repo {
 						p.Spring,
 						p.Summer,
 						p.Winter,
-						p.PerfumeTags.Select(tag => new TagDTO(tag.Tag.TagName, tag.Tag.Color, tag.Tag.Id)).ToList()
+						p.PerfumeTags.Select(tag => new TagDto(tag.Tag.TagName, tag.Tag.Color, tag.Tag.Id)).ToList()
 					  ),
 					  p.PerfumeWorns.Any() ? p.PerfumeWorns.Count : 0,
 					  p.PerfumeWorns.Any() ? p.PerfumeWorns.Max(x => x.Created_At) : null
@@ -63,27 +63,27 @@ namespace PerfumeTrackerAPI.Repo {
 				.AsNoTracking()
 				.FirstOrDefaultAsync();
 		}
-		public async Task<PerfumeStatDTO> GetPerfumeStats() {
+		public async Task<PerfumeStatDto> GetPerfumeStats() {
 			var raw = await context
 				.Perfumes
 				.GroupBy(g => 1)
-				.Select(x => new PerfumeStatDTO(
+				.Select(x => new PerfumeStatDto(
 					x.Sum(s => s.Ml),
 					x.Sum(s => s.PerfumeWorns.Count),
 					x.Count()))
 				.ToListAsync();
 			if (raw != null && raw.Count > 0) return raw[0];
-			return new PerfumeStatDTO(0, 0, 0);
+			return new PerfumeStatDto(0, 0, 0);
 		}
-		public record PerfumeResult(ResultTypes ResultType, PerfumeDTO? Perfume = null, string ErrorMsg = null);
-		public async Task<PerfumeResult> AddPerfume(PerfumeDTO dto) {
+		public record PerfumeResult(ResultTypes ResultType, PerfumeDto? Perfume = null, string ErrorMsg = null);
+		public async Task<PerfumeResult> AddPerfume(PerfumeDto Dto) {
 			try {
-				var perfume = dto.Adapt<Perfume>();
+				var perfume = Dto.Adapt<Perfume>();
 				if (perfume == null) return new PerfumeResult(ResultTypes.BadRequest);
 				perfume.Created_At = DateTime.UtcNow;
 				context.Perfumes.Add(perfume);
 				await context.SaveChangesAsync();
-				foreach (var tag in dto.Tags) {
+				foreach (var tag in Dto.Tags) {
 					context.PerfumeTags.Add(new PerfumeTag() {
 						Created_At = DateTime.UtcNow,
 						PerfumeId = perfume.Id,
@@ -91,7 +91,7 @@ namespace PerfumeTrackerAPI.Repo {
 					});
 				}
 				await context.SaveChangesAsync();
-				return new PerfumeResult(ResultTypes.Ok, perfume.Adapt<PerfumeDTO>());
+				return new PerfumeResult(ResultTypes.Ok, perfume.Adapt<PerfumeDto>());
 			}
 			catch (Exception ex) {
 				return new PerfumeResult(ResultTypes.BadRequest, null, ex.Message);
@@ -104,8 +104,8 @@ namespace PerfumeTrackerAPI.Repo {
 			await context.SaveChangesAsync();
 			return new PerfumeResult(ResultTypes.Ok);
 		}
-		public async Task<PerfumeResult> UpdatePerfume(int id, PerfumeDTO dto) {
-			var perfume = dto.Adapt<Perfume>();
+		public async Task<PerfumeResult> UpdatePerfume(int id, PerfumeDto Dto) {
+			var perfume = Dto.Adapt<Perfume>();
 			if (perfume == null || id != perfume.Id) {
 				return new PerfumeResult(ResultTypes.BadRequest);
 			}
@@ -120,29 +120,30 @@ namespace PerfumeTrackerAPI.Repo {
 			context.Entry(find).CurrentValues.SetValues(perfume);
 			find.Updated_At = DateTime.UtcNow;
 
-			await UpdateTags(dto, find);
+			await UpdateTags(Dto, find);
 
-			return new PerfumeResult(ResultTypes.Ok, find.Adapt<PerfumeDTO>());
+			return new PerfumeResult(ResultTypes.Ok, find.Adapt<PerfumeDto>());
 		}
 
-		public async Task<PerfumeResult> UpdatePerfumeImageGuid(ImageGuidDTO dto) {
-			var find = await context.Perfumes.FindAsync(dto.ParentObjectId);
+		public async Task<PerfumeResult> UpdatePerfumeImageGuid(ImageGuidDto Dto) {
+			var find = await context.Perfumes.FindAsync(Dto.ParentObjectId);
 			if (find == null) return new PerfumeResult(ResultTypes.NotFound);
-			find.ImageObjectKey = dto.ImageObjectKey;
+			find.ImageObjectKey = Dto.ImageObjectKey;
 			find.Updated_At = DateTime.UtcNow;
 			await context.SaveChangesAsync();
-			return new PerfumeResult(ResultTypes.Ok, find.Adapt<PerfumeDTO>());
+			return new PerfumeResult(ResultTypes.Ok, find.Adapt<PerfumeDto>());
 		}
 
-		private async Task UpdateTags(PerfumeDTO dto, Perfume? find) {
+		private async Task UpdateTags(PerfumeDto Dto, Perfume? find) {
+			if (find == null) return;
 			var tagsInDB = find.PerfumeTags
 				.Select(x => x.Tag)
 				.Select(x => x.TagName)
 				.ToList();
-			foreach (var remove in find.PerfumeTags.Where(x => !dto.Tags.Select(x => x.TagName).Contains(x.Tag.TagName))) {
+			foreach (var remove in find.PerfumeTags.Where(x => !Dto.Tags.Select(x => x.TagName).Contains(x.Tag.TagName))) {
 				context.PerfumeTags.Remove(remove);
 			}
-			foreach (var add in dto.Tags.Where(x => !tagsInDB.Contains(x.TagName))) {
+			foreach (var add in Dto.Tags.Where(x => !tagsInDB.Contains(x.TagName))) {
 				context.PerfumeTags.Add(new PerfumeTag() {
 					PerfumeId = find.Id,
 					TagId = add.Id //TODO: this is not good, tag ID is coming back from client side
